Cambios en todos los servicios para mantener consistencia en el código y protegidos todos los endpoints sensibles para administradores

+++ backend/service_productos/apps/productos/views.py
@@ -5,10 +5,11 @@ from rest_framework import status
 from .models import Producto
 from .serializer import ProductoSerializer
 from rest_framework.permissions import IsAuthenticated
-from django.shortcuts import get_object_or_404
+from utils.permissions import IsSuperUser
+from rest_framework.generics import ListAPIView
 
 class ProductoCrearView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsSuperUser]
 
     def post(self, request):
         serializer = ProductoSerializer(data=request.data)
@@ -18,36 +19,66 @@ class ProductoCrearView(APIView):
         return Response(serializer.errors, status=status.HTTP_404_NOT_FOUND)
 
 class ProductoEditarView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsSuperUser]
 
-    def put(self, request, pk):
-        producto = get_object_or_404(Producto, pk=pk)
-        serializer = ProductoSerializer(producto, data=request.data)
-        if serializer.is_valid():
-            serializer.save()
-            return Response(serializer.data, status=status.HTTP_200_OK)
-        return Response(serializer.errors, status=status.HTTP_404_NOT_FOUND)
+    def put(self, request):
+        id = request.query_params.get('id')
+
+        if not id:
+            return Response({'detail':'Falta proporcionar el id del producto a editar'}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            producto = Producto.objects.get(id=id)
+            productoSerializer = ProductoSerializer(producto, data=request.data)
+
+            if productoSerializer.is_valid():
+                return Response({'detail':'Producto editado exitosamente'}, status=status.HTTP_200_OK)
+            else:
+                return Response(productoSerializer.errors, status=status.HTTP_400_BAD_REQUEST)
+        except:
+            return Response({'detail':'Producto a editar no encontrado'}, status=status.HTTP_404_NOT_FOUND)
 
 class ProductoEliminarView(APIView):
-    permission_classes = [IsAuthenticated]
+    permission_classes = [IsAuthenticated, IsSuperUser]
 
     def post(self, request, pk):
-        producto = get_object_or_404(Producto, pk=pk)
-        producto.delete()
-        return Response({'mensaje': 'Producto eliminado correctamente.'}, status=status.HTTP_200_OK)
+        id = request.query_params.get('id')
 
+        if not id:
+            return Response({'detail':'Falta proporcionar el id del producto a eliminar'}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            producto = Producto.objects.get(id=id)
+            producto.delete()
+            return Response({'detail':'Producto eliminado exitosamente'}, status=status.HTTP_200_OK)
+        except:
+            return Response({'detail':'Producto a eliminar no encontrado'}, status=status.HTTP_400_BAD_REQUEST)
+        
 class ProductoListarView(APIView):
     permission_classes = [IsAuthenticated]
 
     def get(self, request):
-        productos = Producto.objects.all()
+        productos = Producto.objects.filter(disponible=True)
         serializer = ProductoSerializer(productos, many=True)
         return Response(serializer.data, status=status.HTTP_200_OK)
 
-class ProductoBuscarView(APIView):
+class ProductoBuscarView(ListAPIView):
+    serializer_class = ProductoSerializer
     permission_classes = [IsAuthenticated]
+    
+    def get_queryset(self):
+        id = self.request.query_params.get('id')
+        nombre = self.request.query_params.get('nombre')
 
-    def get(self, request, pk):
-        producto = get_object_or_404(Producto, pk=pk)
-        serializer = ProductoSerializer(producto)
-        return Response(serializer.data, status=status.HTTP_200_OK)
+        if not id and not nombre:
+            return Response({'detail':'Falta proporcionar al menos id ó nombre del producto a buscar'}, status=status.HTTP_400_BAD_REQUEST)
+
+        try:
+            if not id:
+                producto = Producto.objects.filter(nombre_pedido=nombre, disponible=True)
+                return producto
+            else:
+                producto = Producto.objects.filter(id=id)
+                return producto
+        except:
+            return Producto.objects.none()
